+++ packages/sentry-client/src/sentry/constants.ts
@@ -1,0 +1,18 @@
+export const POLLING_TIME = 5000;
+
+export enum STATUS {
+    OUTAGE,
+    AVAILABLE,
+    UNKNOWN,
+    ISSUE,
+    MAINTENANCE
+}
+
+interface IView {
+    [key: string]: string;
+}
+
+export const view: IView = {
+    servers: "Servers",
+    services: "Services"
+};
