--- packages/sentry-client/src/sentry/constants/polling.ts
@@ -1,1 +1,0 @@
-export const POLLING_TIME = 5000;
