--- packages/sentry-client/src/sentry/constants/status.ts
@@ -1,7 +1,0 @@
-export enum STATUS {
-    OUTAGE,
-    AVAILABLE,
-    UNKNOWN,
-    ISSUE,
-    MAINTENANCE
-}
