+++ packages/sentry-client/test/sentry/components/Status.test.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import * as renderer from "react-test-renderer";
 import { Status } from "../../../src/sentry/components/Status";
-import { STATUS } from "../../../src/sentry/constants/status";
+import { STATUS } from "../../../src/sentry/constants";
 
 describe("Status", () => {
     it("renders an available status", () => {
